+++ CHANGES.rst
@@ -7,6 +7,7 @@ Unreleased
 - Confirm support for `Django 4.0`
 - Add Spanish translation
 - Add French translation
+- Fix AutoLastModifiedField cache (GH-#520)
 
 4.2.0 (2021-10-11)
 ------------------